+++ src/main/kotlin/client/api/cards/CardsController.kt
@@ -20,7 +20,7 @@ class CardsController {
     @ApiOperation(value = "Return card")
     @ApiResponses(value = [ApiResponse(code = 200, message = "Ok")])
     @GetMapping("/cards/{id}")
-    suspend fun getCardsById(@PathVariable id: Int): CardBody {
+    suspend fun getCardsById(@PathVariable id: Long): CardBody {
         log.info("GET Response: /cards/${id}")
         return Card(id).data
     }
@@ -28,7 +28,7 @@ class CardsController {
     @ApiOperation(value = "Return card history")
     @ApiResponses(value = [ApiResponse(code = 200, message = "Ok")])
     @GetMapping("/cards/{id}/history")
-    suspend fun getCardsHistoryById(@PathVariable id: Int): Any {
+    suspend fun getCardsHistoryById(@PathVariable id: Long): Any {
         log.info("GET Response: /cards/${id}/history")
         return Card(id).history
     }
@@ -36,7 +36,7 @@ class CardsController {
     @ApiOperation(value = "Add card")
     @ApiResponses(value = [ApiResponse(code = 200, message = "Ok")])
     @PostMapping("/cards")
-    suspend fun postCards(@RequestBody cardBody: CardBody): Int {
+    suspend fun postCards(@RequestBody cardBody: CardBody): Long {
         log.info("POST Response: /cards")
         return Card().post(cardBody)
     }
@@ -44,7 +44,7 @@ class CardsController {
     @ApiOperation(value = "Card payment")
     @ApiResponses(value = [ApiResponse(code = 200, message = "Ok")])
     @PostMapping("/cards/{id}/payment")
-    suspend fun postCardsPaymentById(@PathVariable id: Int, @RequestBody paymentBody: PaymentBody): Any {
+    suspend fun postCardsPaymentById(@PathVariable id: Long, @RequestBody paymentBody: PaymentBody): Any {
         log.info("POST Response: /cards/${id}/payment")
         return Card(id).pay(paymentBody)
     }
@@ -52,7 +52,7 @@ class CardsController {
     @ApiOperation(value = "Card transfer")
     @ApiResponses(value = [ApiResponse(code = 200, message = "Ok")])
     @PostMapping("/cards/{id}/transfer")
-    suspend fun postCardsTransferById(@PathVariable id: Int, @RequestBody transferBody: TransferBody): Any {
+    suspend fun postCardsTransferById(@PathVariable id: Long, @RequestBody transferBody: TransferBody): Any {
         log.info("POST Response: /cards/${id}/transfer")
         return Card(id).transfer(transferBody)
     }
@@ -60,7 +60,7 @@ class CardsController {
     @ApiOperation(value = "Card receipt")
     @ApiResponses(value = [ApiResponse(code = 200, message = "Ok")])
     @PostMapping("/cards/{id}/receipt")
-    suspend fun postCardsReceiptById(@PathVariable id: Int, @RequestBody transferBody: TransferBody): Any {
+    suspend fun postCardsReceiptById(@PathVariable id: Long, @RequestBody transferBody: TransferBody): Any {
         log.info("POST Response: /cards/${id}/receipt")
         return Card(id).receipt(transferBody)
     }
@@ -68,7 +68,7 @@ class CardsController {
     @ApiOperation(value = "Delete card")
     @ApiResponses(value = [ApiResponse(code = 200, message = "Ok")])
     @DeleteMapping("/cards/{id}")
-    suspend fun deleteCardsById(@PathVariable id: Int) {
+    suspend fun deleteCardsById(@PathVariable id: Long) {
         log.info("DELETE Response: /cards/${id}")
         return Card(id).delete()
     }
